+++ src/Differences.Model/EntityModels/Reputation/UserReputationLog.cs
@@ -12,7 +12,7 @@ namespace Differences.Interaction.EntityModels
         [ExcludeFromCodeCoverage]
         public UserReputationLog() { }
 
-        public UserReputationLog(Guid userId, int reputationTypeId, double value, int? subjectId)
+        public UserReputationLog(Guid userId, int reputationTypeId, decimal value, int? subjectId)
             : this()
         {
             UserId = userId;
@@ -28,7 +28,7 @@ namespace Differences.Interaction.EntityModels
         [Required]
         public int ReputationTypeId { get; private set; }
         [Required]
-        public double Value { get; private set; }
+        public decimal Value { get; private set; }
         public int? SubjectId { get; private set; }
     }
 }
